refactor: wire up the graphql backend API
This change makes the initial dishes and ingredients data coming from
the new real backend GraphQL API instead of reading from local json
files. The ingredients API needs to be updated (see FIXME in the code)
to be more useful though. The change also ensures the UI won't break
when some pieces of data are null. The only constraint is that we
require at least one dish is returned from the API so we can generate
a menu. An error will be thrown if the returned dish data is empty.

The change doesn't cover "share" APIs yet.