refactor: wire up ShareableMenu API for share feature
This change is the corresponding client implementation to use backend
share menu APIs for share feature. We remove the short url to full url
redirect as it was a hack when it was a pure client app. We now have the
full payload stored in the backend so we can query it anytime with a key.

Downgraded pako library version to 1.0 as it seems the latest version
(2.0) has removed to: 'string' option from deflate().